†修正† create_tweet_vote_embedを別のモジュールに分けた
