+++ lib/data/tweet_vote.py
@@ -4,7 +4,6 @@ tweet_vote.py
 ツイートの投票を司るクラスが入っている。
 """
 import math
-import random
 
 import discord
 
@@ -58,43 +57,4 @@ class TweetVote:
         :return: 文字列と化したTweetVote
         """
         return "{}\nAuthored by: {}".format(self.content, self.author)
-
-
-def create_tweet_vote_embed(tweet_vote: TweetVote) -> discord.Embed:
-    """
-    TweetVoteをDiscordのメッセージに埋め込める形に変換する。
-    :return: Embed。
-    """
-
-    # 可決率を計算する。
-    approve_rate = tweet_vote.get_approval_rate()
-
-    # embedを組み立てる
-    embed = discord.Embed()
-    embed.title = "†ツイート審議待ち†"
-    embed.description = "以下のリアクションから投票してください。\nリアクション全部消したらメス堕ちさせるからな"
-    embed.colour = (random.randint(127, 255) << 16) + (random.randint(127, 255) << 8) + random.randint(127, 255)
-
-    embed.add_field(
-        name="ツイート内容",
-        value="```{}```".format(tweet_vote.content),
-        inline=False
-    )
-    embed.add_field(
-        name="ツイートしたい人",
-        value="{} (`{}`)".format(tweet_vote.author.display_name, tweet_vote.author.name),
-        inline=False
-    )
-    embed.add_field(
-        name="投票状況",
-        value=":thumbsup: {}/{} :thumbsdown: ({}%)".format(
-            tweet_vote.approves,
-            tweet_vote.denys,
-            approve_rate
-        ),
-        inline=False
-    )
-
-    # 返す。
-    return embed
