+++ lib/discord/op/command/impl/create_vote_command.py
@@ -10,7 +10,8 @@ from math import ceil
 
 import discord
 
-from lib.data.tweet_vote import TweetVote, create_tweet_vote_embed
+from lib.data.tweet_vote import TweetVote
+from lib.data.tweet_vote_utils import create_tweet_vote_embed
 from lib.data.tweet_votes_record import TweetsVoteRecord
 from lib.discord.op.command.abst_command_base import AbstCommandBase
 from lib.discord.op.command.command_property import CommandProperty
