+++ lib/discord/op/event/reaction_event.py
@@ -5,7 +5,7 @@ reaction_event.py
 """
 import discord
 
-from lib.data.tweet_vote import create_tweet_vote_embed
+from lib.data.tweet_vote_utils import create_tweet_vote_embed
 from lib.data.tweet_votes_record import TweetsVoteRecord
 from lib.logging.logger import log
 from lib.settings.setting import Setting
